smart-scheduler/requirements.txt
google-adk==1.3.0
google-genai==1.20.0
websockets>=13.0.0,<14.0.0
google-api-python-client>=2.157.0
google-auth-oauthlib>=1.0.0
pytz
python-dotenv>=1.0.0 